+++ frontend/src/Components/Layout/index.js
@@ -1,0 +1,3 @@
+import LayoutContainer from "./LayoutContainer";
+
+export default LayoutContainer;
