+++ frontend/src/Components/Main/MainContainer.js
@@ -1,0 +1,16 @@
+import React, { useState } from "react";
+import MainPresenter from "./MainPresenter";
+
+const MainContainer = () => {
+  const [searchKeyword, setSearchKeyword] = useState("");
+
+  const handleChange = (event) => {
+    setSearchKeyword(event.target.value);
+  };
+
+  return (
+    <MainPresenter searchKeyword={searchKeyword} handleChange={handleChange} />
+  );
+};
+
+export default MainContainer;
