--- frontend/src/Components/Main/MainContent/index.js
@@ -1,3 +1,0 @@
-import MainContentContainer from "./MainContentContainer";
-
-export default MainContentContainer;
