+++ frontend/src/Components/Main/index.js
@@ -1,0 +1,3 @@
+import MainContainer from "./MainContainer";
+
+export default MainContainer;
