+++ frontend/src/Pages/LoginPage.js
@@ -1,22 +1,14 @@
-import React, { useState } from "react";
+import React from "react";
 
 import Login from "../Components/Login";
-import SideMenu from "../Components/Main/SideMenu";
-import SideMenuBtn from "../Components/Main/SideMenuBtn";
+import Layout from "../Components/Layout";
 import "../scss/Login.scss";
 
 const LoginPage = () => {
-  const [showSideMenu, setShowSideMenu] = useState(true);
-
   return (
-    <>
-      <SideMenuBtn
-        showSideMenu={showSideMenu}
-        setShowSideMenu={setShowSideMenu}
-      />
-      {showSideMenu ? <SideMenu /> : null}
+    <Layout>
       <Login />
-    </>
+    </Layout>
   );
 };
