+++ frontend/src/scss/Login.scss
@@ -6,4 +6,4 @@
 @import './base/mixin';
 
 // components scss
-@import './components/_login';
+@import './components/login';
