+++ frontend/src/scss/MainPage.scss
@@ -7,5 +7,4 @@
 
 // components scss
 @import "./components/mainContent";
-@import "./components/mainSideMenu";
-@import "./components/sideMenuBtn";
+@import "./Layout";
