+++ frontend/src/scss/components/_login.scss
@@ -1,6 +1,6 @@
 .login-container {
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
+  height: 100%;
   display: flex;
   flex-direction: column;
   align-items: center;
